+++ packages/in-memory/.size-snapshot.json
@@ -1,8 +1,8 @@
 {
   "dist/hickory-in-memory.es.js": {
-    "bundled": 4439,
-    "minified": 1422,
-    "gzipped": 631,
+    "bundled": 5323,
+    "minified": 1786,
+    "gzipped": 706,
     "treeshaked": {
       "rollup": {
         "code": 22,
@@ -14,18 +14,18 @@
     }
   },
   "dist/hickory-in-memory.js": {
-    "bundled": 4503,
-    "minified": 1481,
-    "gzipped": 681
+    "bundled": 5415,
+    "minified": 1871,
+    "gzipped": 759
   },
   "dist/hickory-in-memory.umd.js": {
-    "bundled": 15202,
-    "minified": 4374,
-    "gzipped": 1856
+    "bundled": 16227,
+    "minified": 4686,
+    "gzipped": 1924
   },
   "dist/hickory-in-memory.min.js": {
-    "bundled": 15202,
-    "minified": 4374,
-    "gzipped": 1856
+    "bundled": 16227,
+    "minified": 4686,
+    "gzipped": 1924
   }
 }
