+++ packages/in-memory/types/InMemory.d.ts
@@ -1,0 +1,3 @@
+import { ResponseHandler } from "@hickory/root";
+import { InMemoryOptions, InMemoryHistory } from "./types";
+export declare function InMemory(fn: ResponseHandler, options?: InMemoryOptions): InMemoryHistory;
