+++ packages/in-memory/types/InMemoryLite.d.ts
@@ -1,0 +1,3 @@
+import { ResponseHandler } from "@hickory/root";
+import { InMemoryLiteOptions, InMemoryLiteHistory } from "./types";
+export declare function InMemoryLite(fn: ResponseHandler, options: InMemoryLiteOptions): InMemoryLiteHistory;
