+++ packages/in-memory/types/index.d.ts
@@ -1,4 +1,4 @@
-import { ResponseHandler } from "@hickory/root";
-import { InMemoryOptions, InMemoryHistory } from "./types";
 export * from "./types";
-export declare function InMemory(fn: ResponseHandler, options?: InMemoryOptions): InMemoryHistory;
+import { InMemory } from "./InMemory";
+import { InMemoryLite } from "./InMemoryLite";
+export { InMemory, InMemoryLite };
